Reorder namespace and global config deletion command
The idea is namespace removal takes the longest, so we'd rather leave it
hanging than config deletion. Then is user gets tired of waiting and
sends a SIGINT we don't leave the global config dangling - just the
namespace removal, which will complete in the background.
So it's just a user experience improvement.

Signed-off-by: MOZGIII <mike-n@narod.ru>